Extract TheFadeCharacterSheet into src/character-sheet.js
Move the ~4,770-line character sheet class out of thefade.js into its
own file. No logic change; class is relocated as-is and re-imported.

thefade.js: 5352 → 558 lines. What remains is the entry imports,
registerItemSheets, and the five system hooks — a proper entry file.

Also prune thefade.js's import list to only the names it actually uses
now that classes have moved (constants and helper functions imported
only by relocated classes no longer need to flow through the entry).

The character-sheet.js class is still a ~4,700-line god-class; the
mixin-based split is a separate follow-up.
